# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Garden Maintenance System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plants/catalog:
    get:
      operationId: listCatalogPlants
      tags: [plants]
      summary: Browse the global plant catalog
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
      responses:
        "200":
          description: Catalog plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogPlant"
  /plants/catalog/{plantId}:
    get:
      operationId: getCatalogPlant
      tags: [plants]
      summary: Get a single catalog plant with full details
      parameters:
        - in: path
          name: plantId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Catalog plant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogPlantDetail"
...
  /garden:
    get:
      operationId: listGarden
      tags: [garden]
      summary: List the current user's plants
      responses:
        "200":
          description: Garden plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GardenPlant"
    post:
      operationId: addGardenPlant
      tags: [garden]
      summary: Add a plant to the user's garden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGardenPlantBody"
      responses:
        "200":
          description: Created garden plant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenPlant"
  /garden/{plantId}:
    get:
      operationId: getGardenPlant
      tags: [garden]
      summary: Get a garden plant with care routine and recent activity
      parameters:
        - in: path
          name: plantId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Garden plant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenPlantDetail"
    patch:
      operationId: updateGardenPlant
      tags: [garden]
      summary: Update a garden plant
      parameters:
        - in: path
          name: plantId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGardenPlantBody"
      responses:
        "200":
          description: Updated garden plant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenPlant"
    delete:
      operationId: deleteGardenPlant
      tags: [garden]
      summary: Remove a plant from the garden
      parameters:
        - in: path
          name: plantId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
...
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List the user's reminders
      parameters:
        - in: query
          name: gardenPlantId
          schema: { type: integer }
        - in: query
          name: upcoming
          schema: { type: boolean }
      responses:
        "200":
          description: Reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderBody"
      responses:
        "200":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{reminderId}:
    patch:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReminderBody"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
  /reminders/{reminderId}/complete:
    post:
      operationId: completeReminder
      tags: [reminders]
      summary: Mark a reminder as completed
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Completed reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
...
components:
  schemas:
...
    SimpleOk:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    CatalogPlant:
      type: object
      required: [id, commonName, scientificName, category, difficulty, imageUrl, summary]
      properties:
        id: { type: integer }
        commonName: { type: string }
        scientificName: { type: string }
        category: { type: string }
        difficulty: { type: string, enum: [easy, medium, hard] }
        imageUrl: { type: string }
        summary: { type: string }
        sunlight: { type: string }
        waterFrequencyDays: { type: integer }
        idealTemperatureC: { type: string }
        soilType: { type: string }
    CatalogPlantDetail:
      allOf:
        - $ref: "#/components/schemas/CatalogPlant"
        - type: object
          required: [description, articles, videos]
          properties:
            description: { type: string }
            articles:
              type: array
              items:
                $ref: "#/components/schemas/Article"
            videos:
              type: array
              items:
                $ref: "#/components/schemas/Video"
    Article:
      type: object
      required: [title, url, source]
      properties:
        title: { type: string }
        url: { type: string }
        source: { type: string }
    Video:
      type: object
      required: [title, url, thumbnailUrl]
      properties:
        title: { type: string }
        url: { type: string }
        thumbnailUrl: { type: string }
        durationSeconds: { type: integer }
...
    GardenPlant:
      type: object
      required: [id, nickname, location, plantedAt, catalogPlant]
      properties:
        id: { type: integer }
        nickname: { type: string }
        location: { type: string }
        plantedAt: { type: string, format: date-time }
        notes: { type: string, nullable: true }
        healthStatus: { type: string, enum: [thriving, healthy, watch, struggling] }
        lastWateredAt: { type: string, format: date-time, nullable: true }
        catalogPlant:
          $ref: "#/components/schemas/CatalogPlant"
    GardenPlantDetail:
      allOf:
        - $ref: "#/components/schemas/GardenPlant"
        - type: object
          properties:
            careRoutine:
              oneOf:
                - $ref: "#/components/schemas/CareRoutine"
                - type: "null"
            recentWatering:
              type: array
              items:
                $ref: "#/components/schemas/WateringLog"
            upcomingReminders:
              type: array
              items:
                $ref: "#/components/schemas/Reminder"
            latestSoil:
              oneOf:
                - $ref: "#/components/schemas/SoilReading"
                - type: "null"
    AddGardenPlantBody:
      type: object
      required: [catalogPlantId, nickname, location]
      properties:
        catalogPlantId: { type: integer }
        nickname: { type: string }
        location: { type: string }
        notes: { type: string }
    UpdateGardenPlantBody:
      type: object
      properties:
        nickname: { type: string }
        location: { type: string }
        notes: { type: string }
        healthStatus: { type: string, enum: [thriving, healthy, watch, struggling] }
    CareRoutine:
      type: object
      required: [id, gardenPlantId, wateringFrequencyDays, wateringAmountMl, sunlight, soilNotes, fertilizerNotes, generatedAt, tasks]
      properties:
        id: { type: integer }
        gardenPlantId: { type: integer }
        wateringFrequencyDays: { type: integer }
        wateringAmountMl: { type: integer }
        sunlight: { type: string }
        soilNotes: { type: string }
        fertilizerNotes: { type: string }
        generatedAt: { type: string, format: date-time }
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/CareTask"
    CareTask:
      type: object
      required: [title, frequency, description]
      properties:
        title: { type: string }
        frequency: { type: string }
        description: { type: string }
...
    Reminder:
      type: object
      required: [id, gardenPlantId, title, type, scheduledAt, completed, repeatDays]
      properties:
        id: { type: integer }
        gardenPlantId: { type: integer }
        gardenPlantNickname: { type: string }
        title: { type: string }
        type: { type: string, enum: [water, fertilize, prune, inspect, other] }
        scheduledAt: { type: string, format: date-time }
        completed: { type: boolean }
        repeatDays: { type: integer }
        notes: { type: string, nullable: true }
    CreateReminderBody:
      type: object
      required: [gardenPlantId, title, type, scheduledAt]
      properties:
        gardenPlantId: { type: integer }
        title: { type: string }
        type: { type: string, enum: [water, fertilize, prune, inspect, other] }
        scheduledAt: { type: string, format: date-time }
        repeatDays: { type: integer }
        notes: { type: string }
    UpdateReminderBody:
      type: object
      properties:
        title: { type: string }
        type: { type: string, enum: [water, fertilize, prune, inspect, other] }
        scheduledAt: { type: string, format: date-time }
        repeatDays: { type: integer }
        notes: { type: string }
        completed: { type: boolean }
    SoilReading:
      type: object
      required: [gardenPlantId, recordedAt, ph, temperatureC, humidityPct, moisturePct, status]
      properties:
        gardenPlantId: { type: integer }
        recordedAt: { type: string, format: date-time }
        ph: { type: number }
        temperatureC: { type: number }
        humidityPct: { type: number }
        moisturePct: { type: number }
        status: { type: string, enum: [optimal, watch, alert] }
    WateringLog:
      type: object
      required: [id, gardenPlantId, gardenPlantNickname, wateredAt, amountMl, source]
      properties:
        id: { type: integer }
        gardenPlantId: { type: integer }
        gardenPlantNickname: { type: string }
        wateredAt: { type: string, format: date-time }
        amountMl: { type: integer }
        source: { type: string, enum: [manual, reminder, auto] }
        notes: { type: string, nullable: true }
...